# openapi.yaml
openapi: 3.1.0
info:
  title: Dash API
  version: 0.1.0
  description: >
    Single source of truth for the Dash API. The TypeScript client in
    packages/api-client is generated from this file — do not hand-write
    API types in the frontend.
servers:
  - url: /
paths:
...
  /api/notify/test:
    post:
      operationId: notifyTest
      summary: Send a test notification through configured transports
      description: >-
        Fires a test event synchronously through every configured notify_*
        transport, or only the transport named in the body. Returns a
        transport -> "ok" | "error: ..." map.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                transport:
                  type: string
                  enum: [webhook, slack, discord, telegram, gotify, ntfy, smtp]
      responses:
        "200":
          description: Per-transport delivery results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: object
                    additionalProperties:
                      type: string
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
...
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...